warning removal

+++ src/gaussian_splatting_ui.cpp
@@ -174,7 +174,7 @@ void GaussianSplatting::onUIRender()
         else
           createDataBuffers();
         createPipelines();
-        m_plyLoader.reset(); // change status to READY
+        m_plyLoader.reset();  // change status to READY
         //
         ImGui::CloseCurrentPopup();
         addToRecentFiles(m_loadedSceneFilename);
@@ -339,11 +339,9 @@ void GaussianSplatting::onUIRender()
       ImGui::BeginDisabled();
 
       PE::entry(
-          "Distances  (ms)",
-          [&]() { return ImGui::InputFloat("##HiddenID", (float*)&(m_distTime), 0, 100000); }, "TODOC");
+          "Distances  (ms)", [&]() { return ImGui::InputInt("##HiddenID", (int*)&(m_distTime), 0, 100000); }, "TODOC");
       PE::entry(
-          "Sorting  (ms)",
-          [&]() { return ImGui::InputFloat("##HiddenID", (float*)&(m_sortTime), 0, 100000); }, "TODOC");
+          "Sorting  (ms)", [&]() { return ImGui::InputInt("##HiddenID", (int*)&(m_sortTime), 0, 100000); }, "TODOC");
       uint32_t totalSplatCount = (uint32_t)gsIndex.size();
       PE::entry(
           "Total splats", [&]() { return ImGui::InputInt("##HiddenID", (int*)&totalSplatCount, 0, 100000); }, "TODOC");
@@ -515,7 +513,6 @@ void GaussianSplatting::onUIRender()
     }
     ImGui::End();
   }
-
 }
 
 void GaussianSplatting::onUIMenu()
